requirements.txt
chainlit
fastapi
uvicorn
openai
pyodbc
jinja2
python-dotenv
